🚚 rename: UserProfile Type 파일 이동
