+++ src/hooks/User/useCreateUserProfile.tsx
@@ -1,5 +1,6 @@
 import { createUserProfile } from "@/api/database/User/user.api";
-import { UserProfile } from "@/types/User/User.type";
+import { UserProfile } from "@/lib/user/type";
+
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 export function useCreateUserProfile(userId: string) {
